open_steam(): Parse URL input
Parse the URL coming in from the TCP stream to check that it's a valid
URL. This allows us to fail early and not try to open strings that are
not URLs. Leverage the "url" crate to do the parsing.

Return a `Box<Error>` from `open_stream()` to enable us to return either
an `io` or a `Url` error in our `Result`.

Remove the `trim_right()` call on `url` since it's now a `Url` type
instead of a `String`, and as a result of parsing to a `Url`, it no
longer has the problematic trailing newline.